Keep Viewer modules within size gates
