+++ crates/oasis7/src/viewer/runtime_live.rs
@@ -91,7 +91,10 @@ pub use config::{ChainLinkPolicy, ViewerRuntimeLiveServerConfig, ViewerRuntimeLi
 use constants::*;
 pub use control_plane::runtime_agent_chat_echo_enabled_from_env;
 use control_plane::{RuntimeLlmSidecar, RuntimePlayerBindingPlan};
-use control_utils::{control_mode_for_action, control_mode_label, runtime_control_error_details};
+use control_utils::{
+    control_mode_for_action, control_mode_label, runtime_control_error_details,
+    should_emit_runtime_advance_snapshot,
+};
 use decision_trace::{append_decision_upstream_trace, decision_trace_provider_error_retryable};
 use gameplay_snapshot::{
     PlayerGameplayCausalitySignal, apply_runtime_snapshot_empty_entities_blocker,
@@ -150,28 +153,7 @@ pub struct ViewerRuntimeLiveServer {
     authoritative_recovery_dir_override: Option<PathBuf>,
 }
 
-fn should_emit_runtime_advance_snapshot(
-    session: &mut RuntimeLiveSession,
-    action: &str,
-    emit_while_paused: bool,
-) -> bool {
-    let is_background_play = action == "play" && session.playing && !emit_while_paused;
-    if !is_background_play {
-        return true;
-    }
-    session.should_emit_background_snapshot(BACKGROUND_PLAY_SNAPSHOT_INTERVAL)
-}
 impl ViewerRuntimeLiveServer {
-    /// Seeds the opt-in local S6 world used to verify smelter affordability UI.
-    pub fn seed_smelter_affordability_debug_scenario(&mut self) -> Result<(), String> {
-        self.seed_smelter_affordability_debug_scenario_inner()
-    }
-
-    /// Seeds the opt-in local S6 world used to verify governance vote quotes.
-    pub fn seed_governance_vote_quote_debug_scenario(&mut self) -> Result<(), String> {
-        self.seed_governance_vote_quote_debug_scenario_inner()
-    }
-
     pub fn new(
         config: ViewerRuntimeLiveServerConfig,
     ) -> Result<Self, ViewerRuntimeLiveServerError> {
